moved ui to ui.py
